# docs/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Search GDELT, MediaCloud, ACLED, The New York Times, The Guardian, and NewsAPI by keyword and date range.",
    "title": "Historical News Search Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/export/csv": {
      "get": {
        "description": "Download the current source page as comma-separated values (CSV).",
        "operationId": "export_csv_api_export_csv_get",
        "parameters": [
          {
            "description": "Keyword or boolean query",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Keyword or boolean query",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "description": "Start date (YYYY-MM-DD)",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "description": "End date (YYYY-MM-DD)",
              "title": "End",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated source names (default: all available)",
            "in": "query",
            "name": "sources",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated source names (default: all available)",
              "title": "Sources",
              "type": "string"
            }
          },
          {
            "description": "Language filter such as 'en', 'english', or 'en-US'. Provider labels are normalized before comparison.",
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "",
              "description": "Language filter such as 'en', 'english', or 'en-US'. Provider labels are normalized before comparison.",
              "title": "Language",
              "type": "string"
            }
          },
          {
            "description": "Collapse duplicate articles that appear across sources",
            "in": "query",
            "name": "dedupe",
            "required": false,
            "schema": {
              "default": true,
              "description": "Collapse duplicate articles that appear across sources",
              "title": "Dedupe",
              "type": "boolean"
            }
          },
          {
            "description": "Optional exact phrase to require in post-filtering",
            "in": "query",
            "name": "exact_phrase",
            "required": false,
            "schema": {
              "default": "",
              "description": "Optional exact phrase to require in post-filtering",
              "title": "Exact Phrase",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated terms to exclude from results",
            "in": "query",
            "name": "exclude_terms",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated terms to exclude from results",
              "title": "Exclude Terms",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated domains to include in local filtering",
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated domains to include in local filtering",
              "title": "Domain",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated domains to exclude from local filtering",
            "in": "query",
            "name": "exclude_domains",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated domains to exclude from local filtering",
              "title": "Exclude Domains",
              "type": "string"
            }
          },
          {
            "description": "Local filter scope: 'all' or 'title'",
            "in": "query",
            "name": "search_scope",
            "required": false,
            "schema": {
              "default": "all",
              "description": "Local filter scope: 'all' or 'title'",
              "title": "Search Scope",
              "type": "string"
            }
          },
          {
            "description": "Keyword post-filter mode: 'provider', 'all_terms', or 'any_term'",
            "in": "query",
            "name": "match_mode",
            "required": false,
            "schema": {
              "default": "provider",
              "description": "Keyword post-filter mode: 'provider', 'all_terms', or 'any_term'",
              "title": "Match Mode",
              "type": "string"
            }
          },
          {
            "description": "Source ranking mode where supported",
            "in": "query",
            "name": "provider_sort",
            "required": false,
            "schema": {
              "default": "default",
              "description": "Source ranking mode where supported",
              "title": "Provider Sort",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated section filters for providers that support them",
            "in": "query",
            "name": "section",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated section filters for providers that support them",
              "title": "Section",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated New York Times news desk filters",
            "in": "query",
            "name": "news_desk",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated New York Times news desk filters",
              "title": "News Desk",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated Guardian tags such as business/economics",
            "in": "query",
            "name": "guardian_tag",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated Guardian tags such as business/economics",
              "title": "Guardian Tag",
              "type": "string"
            }
          },
          {
            "description": "NewsAPI field scope for q: all, title, description, content",
            "in": "query",
            "name": "newsapi_search_in",
            "required": false,
            "schema": {
              "default": "all",
              "description": "NewsAPI field scope for q: all, title, description, content",
              "title": "Newsapi Search In",
              "type": "string"
            }
          },
          {
            "description": "Sort order: 'date_desc' or 'date_asc'",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "date_desc",
              "description": "Sort order: 'date_desc' or 'date_asc'",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "1-based source page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based source page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Csv"
      }
    },
    "/api/export/json": {
      "get": {
        "description": "Download the current source page as JavaScript Object Notation (JSON).",
        "operationId": "export_json_api_export_json_get",
        "parameters": [
          {
            "description": "Keyword or boolean query",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Keyword or boolean query",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "description": "Start date (YYYY-MM-DD)",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "description": "End date (YYYY-MM-DD)",
              "title": "End",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated source names (default: all available)",
            "in": "query",
            "name": "sources",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated source names (default: all available)",
              "title": "Sources",
              "type": "string"
            }
          },
          {
            "description": "Language filter such as 'en', 'english', or 'en-US'. Provider labels are normalized before comparison.",
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "",
              "description": "Language filter such as 'en', 'english', or 'en-US'. Provider labels are normalized before comparison.",
              "title": "Language",
              "type": "string"
            }
          },
          {
            "description": "Collapse duplicate articles that appear across sources",
            "in": "query",
            "name": "dedupe",
            "required": false,
            "schema": {
              "default": true,
              "description": "Collapse duplicate articles that appear across sources",
              "title": "Dedupe",
              "type": "boolean"
            }
          },
          {
            "description": "Optional exact phrase to require in post-filtering",
            "in": "query",
            "name": "exact_phrase",
            "required": false,
            "schema": {
              "default": "",
              "description": "Optional exact phrase to require in post-filtering",
              "title": "Exact Phrase",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated terms to exclude from results",
            "in": "query",
            "name": "exclude_terms",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated terms to exclude from results",
              "title": "Exclude Terms",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated domains to include in local filtering",
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated domains to include in local filtering",
              "title": "Domain",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated domains to exclude from local filtering",
            "in": "query",
            "name": "exclude_domains",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated domains to exclude from local filtering",
              "title": "Exclude Domains",
              "type": "string"
            }
          },
          {
            "description": "Local filter scope: 'all' or 'title'",
            "in": "query",
            "name": "search_scope",
            "required": false,
            "schema": {
              "default": "all",
              "description": "Local filter scope: 'all' or 'title'",
              "title": "Search Scope",
              "type": "string"
            }
          },
          {
            "description": "Keyword post-filter mode: 'provider', 'all_terms', or 'any_term'",
            "in": "query",
            "name": "match_mode",
            "required": false,
            "schema": {
              "default": "provider",
              "description": "Keyword post-filter mode: 'provider', 'all_terms', or 'any_term'",
              "title": "Match Mode",
              "type": "string"
            }
          },
          {
            "description": "Source ranking mode where supported",
            "in": "query",
            "name": "provider_sort",
            "required": false,
            "schema": {
              "default": "default",
              "description": "Source ranking mode where supported",
              "title": "Provider Sort",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated section filters for providers that support them",
            "in": "query",
            "name": "section",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated section filters for providers that support them",
              "title": "Section",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated New York Times news desk filters",
            "in": "query",
            "name": "news_desk",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated New York Times news desk filters",
              "title": "News Desk",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated Guardian tags such as business/economics",
            "in": "query",
            "name": "guardian_tag",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-separated Guardian tags such as business/economics",
              "title": "Guardian Tag",
              "type": "string"
            }
          },
          {
            "description": "NewsAPI field scope for q: all, title, description, content",
            "in": "query",
            "name": "newsapi_search_in",
            "required": false,
            "schema": {
              "default": "all",
              "description": "NewsAPI field scope for q: all, title, description, content",
              "title": "Newsapi Search In",
              "type": "string"
            }
          },
          {
            "description": "Sort order: 'date_desc' or 'date_asc'",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "date_desc",
              "description": "Sort order: 'date_desc' or 'date_asc'",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "1-based source page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based source page number",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Json"
      }
    },
...
  }
}