# src/news/api/app.py
"""Build the FastAPI application and serve the browser client.

The application loads credentials and validated settings once, creates the
process-local cache, and passes those objects to the route functions.
"""

from __future__ import annotations

import argparse
import os
from collections.abc import Callable
from pathlib import Path

from dotenv import load_dotenv
from fastapi import Depends, FastAPI, HTTPException
from fastapi.responses import FileResponse, Response
from fastapi.staticfiles import StaticFiles

from news.api.models import (
    FrontendConfigResponse,
    SearchResponse,
    SourceStatusResponse,
)
from news.api.params import SearchQueryParams
from news.exports.formats import format_csv, format_json
from news.search import SearchExecutor, run_search
from news.search.cache import SearchResultCache, build_search_cache
from news.search.errors import SearchValidationError
from news.search.models import SearchResult
from news.sources import get_source_status, search_all_detailed
from news.web.config import AppSettings, load_settings
from news.web.paths import CONFIG_ENVIRONMENT_VARIABLE, env_path, static_dir

APP_TITLE = "Historical News Search Engine"
APP_DESCRIPTION = (
    "Search GDELT, MediaCloud, ACLED, The New York Times, The Guardian, "
    "and NewsAPI by keyword and date range."
)
APP_VERSION = "0.1.0"
SERVER_HOST = "127.0.0.1"
SERVER_PORT = 8000
SourceStatusProvider = Callable[[], list[dict[str, object]]]


def create_app(
    settings: AppSettings,
    *,
    search_cache: SearchResultCache | None = None,
    search_executor: SearchExecutor = search_all_detailed,
    source_status_provider: SourceStatusProvider = get_source_status,
) -> FastAPI:
    """Build an application from validated settings and supplied dependencies.

    Parameters
    ----------
    settings : AppSettings
        Validated browser and cache settings.
    search_cache : SearchResultCache | None, optional
        Cache supplied by a caller or test. ``None`` builds one from the settings.
    search_executor : SearchExecutor, optional
        Function that queries sources. Tests may supply an offline fake.
    source_status_provider : SourceStatusProvider, optional
        Function that reports source status. Tests may supply deterministic data.

    Returns
    -------
    FastAPI
        Application with routes and packaged browser files.
    """
    active_cache = (
        build_search_cache(settings.cache) if search_cache is None else search_cache
    )
    static_assets = static_dir()
    application = FastAPI(
        title=APP_TITLE,
        description=APP_DESCRIPTION,
        version=APP_VERSION,
    )
    application.state.settings = settings
    application.state.search_cache = active_cache
    application.mount(
        "/static",
        StaticFiles(directory=str(static_assets)),
        name="static",
    )

    @application.get("/")
    async def index() -> FileResponse:
        """Serve the browser app."""
        return FileResponse(str(static_assets / "index.html"))

    @application.get("/api/config", response_model=FrontendConfigResponse)
    async def config() -> dict[str, object]:
        """Return the validated browser settings."""
        return settings.frontend.to_dict()

    @application.get("/api/sources", response_model=list[SourceStatusResponse])
    async def sources() -> list[dict[str, object]]:
        """Return source descriptions and availability."""
        return source_status_provider()

    @application.get("/api/search", response_model=SearchResponse)
    async def search(params: SearchQueryParams = Depends()) -> dict[str, object]:
        """Search the selected sources and return one merged page."""
        result = await _run_search_request(
            params,
            cache=active_cache,
            executor=search_executor,
        )
        return result.to_payload()

    @application.get("/api/export/csv")
    async def export_csv(params: SearchQueryParams = Depends()) -> Response:
        """Download the current source page as comma-separated values (CSV)."""
        result = await _run_search_request(
            params,
            cache=active_cache,
            executor=search_executor,
        )
        return Response(
            content=format_csv(result.articles),
            media_type="text/csv",
            headers={"Content-Disposition": 'attachment; filename="news_export.csv"'},
        )

    @application.get("/api/export/json")
    async def export_json(params: SearchQueryParams = Depends()) -> Response:
        """Download the current source page as JavaScript Object Notation (JSON)."""
        result = await _run_search_request(
            params,
            cache=active_cache,
            executor=search_executor,
        )
        return Response(
            content=format_json(result.articles),
            media_type="application/json",
            headers={"Content-Disposition": 'attachment; filename="news_export.json"'},
        )

    return application


def create_configured_app(
    config_file: Path | str | None = None,
) -> FastAPI:
    """Load local credentials and settings, then construct the application.

    Parameters
    ----------
    config_file : Path | str | None, optional
        Explicit TOML configuration path. ``None`` applies normal resolution.

    Returns
    -------
    FastAPI
        Application configured for the current process.
    """
    load_dotenv(env_path())
    return create_app(load_settings(config_file))


async def _run_search_request(
    params: SearchQueryParams,
    *,
    cache: SearchResultCache,
    executor: SearchExecutor,
) -> SearchResult:
    """Validate request parameters and run the common search process."""
    try:
        request = params.to_search_request()
    except SearchValidationError as exc:
        raise HTTPException(status_code=422, detail=exc.message) from exc
    return await run_search(request, executor=executor, cache=cache)


def main(argv: list[str] | None = None) -> None:
    """Start the local FastAPI server.

    Parameters
    ----------
    argv : list[str] | None, optional
        Command arguments. ``None`` reads the process arguments.
    """
    import uvicorn

    parser = argparse.ArgumentParser(description="Run the news search server.")
    parser.add_argument(
        "--config",
        help=(
            "TOML settings path. Overrides NEWS_CONFIG and the "
            "current-directory config.toml."
        ),
    )
    parser.add_argument(
        "--host",
        default=SERVER_HOST,
        help=f"Interface on which to listen (default: {SERVER_HOST})",
    )
    parser.add_argument(
        "--port",
        type=int,
        default=SERVER_PORT,
        help=f"TCP port on which to listen (default: {SERVER_PORT})",
    )
    parser.add_argument(
        "--reload",
        action="store_true",
        help="Restart the development server when source files change",
    )
    args = parser.parse_args(argv)
    if args.config:
        os.environ[CONFIG_ENVIRONMENT_VARIABLE] = args.config

    uvicorn.run(
        "news.api.app:create_configured_app",
        factory=True,
        host=args.host,
        port=args.port,
        reload=args.reload,
    )


if __name__ == "__main__":
    main()